requirements.txt
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
attrs==25.4.0
bitsandbytes==0.48.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
datasets==4.4.1
deepspeed==0.18.2
diffusers @ git+https://github.com/huggingface/diffusers@b7df4a5387fd65edf4f16fd5fc8c87a7c815a4c7
dill==0.4.0
einops==0.8.1
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.2.0
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.18
networkx==3.5
ninja==1.13.0
numpy==2.3.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
optimum==2.0.0
optimum-quanto==0.2.7
packaging==25.0
pandas==2.3.3
peft==0.18.0
pillow==12.0.0
platformdirs==4.5.0
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
py-cpuinfo==9.0.0
pyarrow==22.0.0
pydantic==2.12.4
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
safetensors==0.6.2
sentry-sdk==2.45.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformer_engine==2.9.0
transformers==4.57.1
triton==3.5.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wandb==0.23.0
wheel==0.45.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
